Fixed a bug when opening non-compressed VCFs.
From SO
http://stackoverflow.com/questions/3703276/how-to-tell-if-a-file-is-gzip-compressed:

"One problem with your approach is, that gzip.GzipFile() will not throw
an exception if you feed it an uncompressed file. Only a later read()
will. This means, that you would probably have to implement some of your
program logic twice. Ugly."

This means that a gzip.open() will always work for a file, regardless of
whether it is gzipped or not. The fix is to try and read() the purported
GZ file, and then seek back to 0 if it worked. Else, fail back to
regular I/O open().