Optimised multiprocessing in script 2.
This prevents having to spawn the process twice, once for creating the
wormtable, and another time for indexing. Indexing step now occurs
inside the first process after the wormtable is created.

Logically the separation of the steps makes sense, but the overhead of
spawning processes can be time-consuming and costly.

This fix also gets rid of the issue of the indexing step using all cores,
although the fix for that is also in place (but commented out). The core
count was not getting passed to add_all_rowid_indexes() from main(),
that's all!